+++ tree-sitter-hird/queries/highlights.scm
@@ -223,6 +223,7 @@
   ","
   ":"
   "."
+  ".."
   "|"
 ] @punctuation.delimiter